pyproject.toml
[project]
name = "jetlend-test-promocode"
version = "0.1.0"
description = ""
authors = [
    {name = "Leshik1305",email = "realleshik1305@gmail.com"}
]
requires-python = ">=3.12"
dependencies = [
    "django (>=6.0.3,<7.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "dotenv (>=0.9.9,<0.10.0)",
    "djangorestframework (>=3.16.1,<4.0.0)",
    "drf-spectacular (>=0.29.0,<0.30.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "black (>=26.3.1,<27.0.0)",
    "isort (>=8.0.1,<9.0.0)"
]

[tool.isort]
profile = "black"
line_length = 88
